requirements.txt
aioredis==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==3.0.2
Django==5.0.3
django-channels==0.7.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-hide==0.0.2
django-minify-html==1.8.0
django-recaptcha==4.0.0
django-template-obfuscator==0.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
hiredis==2.3.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
minify_html==0.15.0
msgpack==1.0.8
mysqlclient==2.2.4
oauthlib==3.2.2
pillow==10.2.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pytz==2024.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
zope.interface==6.2

gunicorn
